refactor: moved genesis to separate file
